This was a tough one! I tried a bunch of different ways, including using matplotlib.
I eventually figured out how to render a short movie of all the frame changes. Turns out this was not efficient.

The right way was to eventually emit what the "bounding box" area was for each frame, and then zoom in on just that frame. In my case it was 10076 frames in. I'm glad I calculated that bc it was the answer to the second part of the question.